requirements.txt
# requirements.txt — Email Spam Detector
# Install: pip install -r requirements.txt

# Core ML
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0

# NLP preprocessing
nltk>=3.8.1

# Image OCR (multimodal spam detection)
Pillow>=10.2.0
pytesseract>=0.3.10

# Web application
streamlit>=1.32.0
